requirements.txt
chromadb>=0.4.18
google.generativeai
tqdm
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-chroma
lark
streamlit
pypdf
python-docx